Safari-611.1.10/LayoutTests/media/video-restricted-no-preload-auto-expected.txt


Test that "no auto-loading restriction" only allows metadata to load in spite of preload='auto'.

EXPECTED (video.preload == 'auto') OK
** setting video.src
RUN(internals.setMediaElementRestrictions(video, 'AutoPreloadingNotPermitted'))

EVENT(loadedmetadata)
EXPECTED (video.readyState < '2') OK
** Pausing...


** 250ms timer fired
EXPECTED (video.readyState < '2') OK
EXPECTED (video.buffered.length == '0') OK
RUN(video.play())

EVENT(canplaythrough)

END OF TEST
